requirements.txt
aiogram==3.*
fastapi==0.110.*
uvicorn[standard]==0.27.*
python-dotenv==1.*
libsql